+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # optics_functions Changelog
 
+## Version x.x.x
+
+- Added:
+  - Chromaticity calculations
+  - Dispersion calculations
+
+
 ## Version 0.1.0
 
 - Added: